Fix up crash protection logic
Sometimes the "This might crash" screen would pop up when editing, or
when loading a fiddle for the first time, or when switching fiddles, all
dependent on timing.

This removes a setTimeout and replaces it with some solid logic around
the loading and unloading of fiddles. The white zone is for immediate
loading and unloading of passengers only. The red zone is for immediate
loading and unloading of passengers only. There is no stopping in the
white zone. No, the white zone is for loading of passengers and there is
no stopping in a RED zone.